+++ admin/tabs/activity/tracking/detail/ActivityDetailModel.ts
@@ -11,8 +11,7 @@ import {GridModel} from '@xh/hoist/cmp/grid';
 import {HoistModel, lookup, managed} from '@xh/hoist/core';
 import {action, computed, makeObservable, observable} from '@xh/hoist/mobx';
 import {ActivityTrackingModel} from '../ActivityTrackingModel';
-import {fmtJson} from '@xh/hoist/format';
-import {timestampReplacer} from '@xh/hoist/utils/datetime';
+import {fmtJson, timestampReplacer} from '@xh/hoist/format';
 
 export class ActivityDetailModel extends HoistModel {
     @lookup(ActivityTrackingModel) activityTrackingModel: ActivityTrackingModel;
